pyproject.toml
[project]
name = "inspect-podman"
version = "0.1.0"
description = "Podman sandbox environment for Inspect AI"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Inspect Podman Maintainers"}
]
dependencies = [
    "inspect_ai>=0.3.163",
    "PyYAML>=6.0",
    "rich>=13.0",
]

[project.entry-points."inspect_ai"]
"podman" = "inspect_podman.providers:podman"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/inspect_podman"]

[dependency-groups]
dev = [
    "pre-commit>=3.7.0",
    "pytest>=8.0.0",
    "ruff>=0.9.3",
]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # flake8
  "D",   # pydocstyle
  "I",   # isort
]
ignore = [
  "D212",  # multi-line-summary-first-line
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D10"]  # undocumented-(public|magic)-*

[tool.ruff.lint.pydocstyle]
convention = "google"
